refactor: extract usePaymentFieldValidation hook to hooks folder
