+++ frontend/src/features/admin-form/create/builder-and-design/BuilderAndDesignDrawer/FieldListDrawer/field-panels/PaymentsInputPanel/VariablePaymentAmountField.tsx
@@ -1,12 +1,11 @@
 import { Controller, RegisterOptions, UseFormReturn } from 'react-hook-form'
 import { FormControl, HStack } from '@chakra-ui/react'
 
+import { usePaymentFieldValidation } from '~hooks/usePaymentFieldValidation'
 import { dollarsToCents } from '~utils/payments'
 import FormErrorMessage from '~components/FormControl/FormErrorMessage'
 import FormLabel from '~components/FormControl/FormLabel'
 import Input from '~components/Input'
-
-import { usePaymentFieldValidation } from '~features/public-form/components/FormPaymentPage/queries'
 
 import { FormPaymentsInput } from './PaymentsInputPanel'
